requirements.txt
astroid==2.0.4
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
cffi==1.11.5
Click==7.0
colorama==0.4.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
Flask==1.0.2
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.1
numpy==1.15.4
opencv-python==3.4.3.18
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
pycparser==2.19
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
torch==0.3.1.post2
torchvision==0.2.1
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.10.11
